Added tfsemb_{parser, config}.py
